Added NavMesh2D
Tileset now has a navmesh that agents can navigate. However, likely due to the small size of the grid and tilemaps, the baked mesh separates itself from the walls by a whole tile! To fix this issue, any tile outside of the walls in the tileset sprite will be marked as walkable using the NavigationModifierTilemap component in the next commit. I highly recommend "coloring in" walls in the Wall tilemap for the best mesh results (and because I said so, it feels nice 👽)!!