Added Enemy Movement
- The one and only enemy; Roamer, can now move!!!
- An acceleration property has been added to enemies
- Enemies initialize their navmeshagent component by passing their scriptable object's values to it, I recommend not overriding the Start method, but if you must, please include base.Start();
⏰ MERGING BRANCH ⏰
Given that the purpose of this branch is near completed, I will look through the code and newly created objects and components to make sure everything is up to code, hehe. After that, I will be merging this branch with the master. Thank you for your attention to this matter.